--- tomcat/src/main/java/org/apache/coyote/http11/HttpMethod.java
@@ -1,6 +1,0 @@
-package org.apache.coyote.http11;
-
-public enum HttpMethod {
-
-    GET
-}
