+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpMethod.java
@@ -1,0 +1,6 @@
+package org.apache.coyote.http11.request;
+
+public enum HttpMethod {
+
+    GET
+}
